Aplicación de arquitectura por N capas, agregado de controlador, negocio, utilidades, dao de Libro
