+++ Biblioteca-Privada/Entidades/IUsuarios.cs
@@ -4,12 +4,12 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Biblioteca_Privada
+namespace Biblioteca_Privada.Entidades
 {
     public interface IUsuario
     {
         void verLibros();
-        Boolean verificarContraseña(string Contraseña);
+        bool verificarContraseña(string Contraseña);
 
     }
 }
