+++ Biblioteca-Privada/Entidades/Libro.cs
