+++ Biblioteca-Privada/Negocio/LibroNegocio.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Biblioteca_Privada.Negocio
+{
+    public class LibroNegocio : ILibroNegocio
+    {
+
+    }
+}
