+++ Biblioteca-Privada/Utilidades/MySqlConfiguration.cs
@@ -1,0 +1,12 @@
+﻿namespace Biblioteca_Privada.Utilidades
+{
+    public class MySQLConfiguration
+    {
+        public string ConnectionString { get; set; }
+
+        public MySQLConfiguration(String connectionString)
+        {
+            ConnectionString = connectionString;
+        }
+    }
+}
