+++ Biblioteca-Privada/appsettings.json
@@ -1,0 +1,13 @@
+﻿{
+    "Logging": {
+        "LogLevel": {
+            "Default": "Information",
+            "Microsoft.AspNetCore": "Warning"
+        }
+    },
+    "AllowedHosts": "*",
+    "ConnectionStrings": {
+        "MySqlConnection": "server=localhost;port=3306;database=prueba;uid=root;password=admin;"
+    }
+}
+
